remove browser stuff
