+++ dev-packages/e2e-tests/test-applications/node-exports-test-app/scripts/consistentExports.ts
@@ -15,7 +15,6 @@ const NODE_EXPORTS_IGNORE = [
   // Probably generated by transpilation, no need to require it
   '__esModule',
   // Only required from the Node package
-  'setNodeAsyncContextStrategy',
   'getDefaultIntegrationsWithoutPerformance',
   'initWithoutDefaultIntegrations',
   'SentryContextManager',
