+++ packages/opentelemetry/src/index.ts
@@ -1,7 +1,18 @@
-export * from './exports';
+export { getScopesFromContext } from './utils/contextData';
 
-// Node-specific exports
+export { enhanceDscWithOpenTelemetryRootSpanName } from './utils/enhanceDscWithOpenTelemetryRootSpanName';
+
+export { getTraceContextForScope } from './trace';
+
+export { setupEventContextTrace } from './setupEventContextTrace';
+
+export { SentryPropagator } from './propagator';
+export { applyOtelSpanData } from './applyOtelSpanData';
+export { backfillStreamedSpanDataFromOtel } from './utils/backfillStreamedSpanData';
+export { SentryTracerProvider } from './tracerProvider';
+
+export { getSentryResource } from './resource';
+
 export { type AsyncLocalStorageLookup } from './asyncLocalStorageContextManager';
 
-// We export the node-specific variant here that uses async local storage
-export { setNodeOpenTelemetryContextAsyncContextStrategy as setOpenTelemetryContextAsyncContextStrategy } from './nodeAsyncContextStrategy';
+export { setOpenTelemetryContextAsyncContextStrategy } from './asyncContextStrategy';
